re organise Alonzon rewards simple stake and tx
